64-bit migration: fix critical type issues and format specifiers
- Use unsigned long for X11 Pixel globals (FGPixel/BGPixel/BDPixel/CRPixel/InvertPixel)
- Fix long→int precision warnings in tfed.c; use int where appropriate, cast pointer diffs
- Correct X11 calls: cast GET_x/GET_y to int, width/height to unsigned int
- Replace %d with %ld/%lx/%p where needed (class.c, cl_menu.c, cl_pane.c, cl_stack.c, event_x.c, cgen.c, hash.c, html.c, fosi.c, tfed.c)
- Fix SWAB macro with __typeof__ to preserve target type
- Sanitize insetHeight handling: use int, avoid double casts
- Clean up various printf formats and pointer prints

Builds clean; warnings reduced and critical 64-bit issues addressed.

+++ src/viola/glib_x.c
@@ -169,7 +169,7 @@ int fontID_symbol_largest = -1;
  */
 Colormap colormap;
 int numOfDisplayColors;
-int FGPixel, BGPixel, BDPixel, CRPixel, InvertPixel;
+unsigned long FGPixel, BGPixel, BDPixel, CRPixel, InvertPixel;
 int DefaultFGPixel, DefaultBGPixel, DefaultBDPixel, DefaultCRPixel;
 int DefaultFGPixel_lighter, DefaultFGPixel_darker, DefaultFGPixel_darkest;
 int DefaultBGPixel_lighter, DefaultBGPixel_darker, DefaultBGPixel_darkest;
